Got web front end auth to work on Mothra by removing session cookie. Instead, we just gen a UUID on service startup, and hide the priv interactions behind that. Login now redirects to that UUID.
Also reworked some web template code to work better in mothra and abaco. Probably not going to spend a lot of time on abaco compatibility otherwise as it doesn’t seem to support drop down menus.

Fixed issue with IP tables not allowing VM to access the webservice on 192.168.9.1 (thank AI chatbot…).

Fixed several issues with Tor not working, such as such as with the optional tor overlay not backing up the torrc properly, as well as messing up the torrc if run multiple times which .

Fixed issue with restoring the clear net not restarting the “find internet” service.